+++ com/baidu/tieba/im/message/ResponseUpdateGroupMessage.java
@@ -22,7 +22,7 @@ public class ResponseUpdateGroupMessage extends SocketResponsedMessage {
     }
 
     /* JADX DEBUG: Method merged with bridge method */
-    @Override // com.baidu.adp.framework.message.b
+    @Override // com.baidu.adp.framework.message.a
     public void decodeInBackGround(int i, byte[] bArr) {
         UpdateGroupResIdl updateGroupResIdl = (UpdateGroupResIdl) new Wire(new Class[0]).parseFrom(bArr, UpdateGroupResIdl.class);
         setError(updateGroupResIdl.error.errorno.intValue());
@@ -42,7 +42,7 @@ public class ResponseUpdateGroupMessage extends SocketResponsedMessage {
                 updateGroupInfoData.setNotice(groupInfo.notice);
                 updateGroupInfoData.setAlbum(groupInfo.album);
                 updateGroupInfoData.setStatus(groupInfo.status.intValue());
-                updateGroupInfoData.setAuthorId(groupInfo.authorId.intValue());
+                updateGroupInfoData.setAuthorId(groupInfo.authorId.longValue());
                 updateGroupInfoData.setAuthorName(groupInfo.authorName);
                 updateGroupInfoData.setCreateTime(groupInfo.createTime.intValue());
                 updateGroupInfoData.setMaxMemberNum(groupInfo.maxMemberNum.intValue());
